Delete api - redux
