+++ app/javascript/redux/actions/actionCreators.js
@@ -1,4 +1,4 @@
-import {LOAD_INTERVIEWS, LOAD_PARTICIPANTS, GET_INTERVIEW, ADD_INTERVIEW, SUBMIT_INTERVIEW, ADD_PARTICIPANT, EDIT_INTERVIEW } from '../actions/actionTypes'
+import {LOAD_INTERVIEWS, LOAD_PARTICIPANTS, GET_INTERVIEW, ADD_INTERVIEW, SUBMIT_INTERVIEW, ADD_PARTICIPANT, EDIT_INTERVIEW, DELETE_INTERVIEW } from '../actions/actionTypes'
 
 export function loadParticipants(participants) {
   return { type: LOAD_PARTICIPANTS, payload: participants }
@@ -28,4 +28,8 @@ export const patchInterview = (data) => {
 
 export const editInterview = (data) => {
   return {type: EDIT_INTERVIEW, payload: data}
+}
+
+export const deleteInterview = (data) => {
+  return {type: DELETE_INTERVIEW, payload: data}
 }
